Add files via upload
Consumer shopping on the internet is growing year by year. However, the conversion
rates have remained more or less the same. For example, most of us browse through
e-commerce websites such as Amazon, perhaps adding items to a wish list or even an
online shopping cart, only to end up buying nothing.
From this common truth comes the need for tools and solutions that can customize
promotions and advertisements for online shoppers and improve this conversion. 

we will be analyzing various factors that affect a purchaser's decision.